Revert "feat: change to spring cloud architecture"
