+++ .vscode/launch.json
@@ -6,13 +6,6 @@
     "configurations": [
         {
             "type": "java",
-            "name": "EurekaServerApplication",
-            "request": "launch",
-            "mainClass": "com.tp.EurekaServerApplication",
-            "projectName": "eureka-server"
-        },
-        {
-            "type": "java",
             "name": "Current File",
             "request": "launch",
             "mainClass": "${file}"
