--- api-gate-service/src/main/java/com/api/gateway/entity/User.java
@@ -1,17 +1,0 @@
-package com.api.gateway.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class User {
-
-    private Long id;
-
-    private String name;
-
-    private String email;
-}
