--- api-gate-service/src/main/java/com/api/gateway/service/HelloService.java
@@ -1,11 +1,0 @@
-package com.api.gateway.service;
-
-import com.api.gateway.vo.HelloRequestVO;
-
-public interface HelloService {
-
-    String echo(String input);
-
-    String hello(HelloRequestVO helloRequestVO);
-}
-
