--- api-gate-service/src/main/java/com/api/gateway/vo/ApiResponse.java
@@ -1,17 +1,0 @@
-package com.api.gateway.vo;
-
-import lombok.Data;
-
-@Data
-public class ApiResponse<T> {
-    private int status;
-    private String message;
-    private T data;
-
-    public ApiResponse(int status, String message, T data) {
-        this.status = status;
-        this.message = message;
-        this.data = data;
-    }
-
-}
