--- api-gate-service/src/test/java/com/api/gateway/ApiGatewayTests.java
@@ -1,13 +1,0 @@
-package com.api.gateway;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApiGatewayTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
