--- eureka-server/src/main/resources/application.yml
@@ -1,7 +1,0 @@
-server:
-  port: 8761
-  
-eureka:
-  client:
-    registerWithEureka: false
-    fetchRegistry: false
