+++ my-app/.editorconfig
