+++ my-app/src/main/java/com/example/demo/entity/User.java
@@ -1,0 +1,17 @@
+package com.example.demo.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class User {
+
+    private Long id;
+
+    private String name;
+
+    private String email;
+}
