+++ my-app/src/main/java/com/example/demo/service/HelloService.java
@@ -1,0 +1,11 @@
+package com.example.demo.service;
+
+import com.example.demo.vo.HelloRequestVO;
+
+public interface HelloService {
+
+    String echo(String input);
+
+    String hello(HelloRequestVO helloRequestVO);
+}
+
