+++ my-app/src/main/java/com/example/demo/vo/ApiResponse.java
@@ -1,0 +1,17 @@
+package com.example.demo.vo;
+
+import lombok.Data;
+
+@Data
+public class ApiResponse<T> {
+    private int status;
+    private String message;
+    private T data;
+
+    public ApiResponse(int status, String message, T data) {
+        this.status = status;
+        this.message = message;
+        this.data = data;
+    }
+
+}
