+++ my-app/src/main/resources/application.yml
@@ -1,0 +1,12 @@
+spring:
+  datasource:
+    url: "jdbc:postgresql://localhost:5432/AlgoCrack"
+    username: postgres
+    password: 32865417
+    driver-class-name: org.postgresql.Driver
+  application:
+    name: demo
+  
+hello:
+  english: Greetings from Spring Boot!
+  chinese: 來自 Spring Boot 的問候！
